+++ system/src/sheets/_module.mjs
@@ -1,5 +1,4 @@
-export {default as FalloutCharacterSheet} from "./FalloutCharacterSheet.mjs";
 export {default as FalloutItemSheet} from "./FalloutItemSheet.mjs";
 export {default as FalloutNpcSheet} from "./FalloutNpcSheet.mjs";
-export {default as FalloutRobotSheet} from "./FalloutRobotSheet.mjs";
+export {default as FalloutPcSheet} from "./FalloutPcSheet.mjs";
 export {default as FalloutSettlementSheet} from "./FalloutSettlementSheet.mjs";